refactor: moved quality_flags into qc
